Works on abstracting toplevel functions to a Utility.kt like readFromJson and rand, working on abstracting functions, previously crowding MainActivity2, to a ScenarioFactory, next is a PawnFactory
