+++ app/src/main/java/com/michaeljahns/namespace/grammy/Scenario.kt
@@ -2,5 +2,5 @@ package com.michaeljahns.namespace.grammy
 
 class Scenario(scenarioLocation: Location, scenarioPawns: MutableList<Pawn>) {
     lateinit var location: Location
-    lateinit var pawnList: MutableList<Pawn>
+    lateinit var pawnList: List<Pawn>
 }
